*: move metric `neogo_version` out of `pkg/network`
Close #3682
